+++ srcs/memory/ft_memset.c
@@ -6,16 +6,16 @@
 /*   By: dcastor <dcastor@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/23 17:14:15 by dcastor           #+#    #+#             */
-/*   Updated: 2025/05/22 11:59:02 by dcastor          ###   ########.fr       */
+/*   Updated: 2025/05/22 14:38:26 by dcastor          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, unsigned int n)
+void	*ft_memset(void *s, int c, size_t n)
 {
 	unsigned char	*dest;
-	unsigned int	i;
+	size_t			i;
 
 	dest = s;
 	i = 0;
